Restructured code and prepared for extensions
Now, we have one source file for each box type.
This file contains all code, related to the box
type. Adding a new box type is thus much simpler:
just copy an existing box type file and adjust it.

Also, we renamed static_plot to page, a much more
natural name, and added the skeleton for a document
object, which will comprise of a set of pages and
can be rendered in one go.

This also allows us to better handle page number,
i.e. we can automatically increase the numbers and
do not have to adjust them when changing e.g. the
ordering of the pages.

These ideas are not yet implemented, but we set
the stage for them with this commit.