Add first support for multi-page documents
This adds multi-page documents. It is still
possible to create a single page, but it is
also possible, to create a document and add
multiple pages to it.

A natural pipeline syntax is used, where each
page is started with start_page, then all
page related statements follow, and then the
next start_page introduces the next page, and
so on.

It turns out, that this is slightly tricky to
implement, since we want the same functions
(e.g. add_plot_box) to work with both a single
page and a document. The solution implemented
here uses two generic functions get_page and
update_page which are specialized for both
document and page. This only slightly complicates
the existing code.