Add support for PDF generation
Using the grDevices::cairo_pdf, we can now
generate both single page slides and multi-page
documents in PDF. The fonts seem to be embedded,
thus the result should be universally usable,
regardless of installed fonts.

It is now possible to save a page either as
PNG - using save_png - or as PDF - using save_pdf.
The whole document can be saved with save_pdf.
It is possible to set a custom filename. If none
is provided, the document title is taken as the
filename.