Support dynamically merged styles
This simplifies the style and theme construction
since it is now possible to merge styles on the
fly everywhere a style is used. If we define base
style modifiers, e.g. "bold" or "right", then we
can adjust a base style by using e.g.
c("bold", "right", "text").

Also renamed document and page function to
create_document and create_page, respectively.
This prevents shadowing of the base "page" function
and increases the clarity of the code.